Refactor client config into modular submodules
