+++ crates/core/src/client/config/builder/arguments.rs
@@ -1,0 +1,31 @@
+use super::*;
+
+impl ClientConfigBuilder {
+    /// Sets the transfer arguments that should be propagated to the engine.
+    #[must_use]
+    pub fn transfer_args<I, S>(mut self, args: I) -> Self
+    where
+        I: IntoIterator<Item = S>,
+        S: Into<OsString>,
+    {
+        self.transfer_args = args.into_iter().map(Into::into).collect();
+        self
+    }
+
+    /// Enables or disables dry-run mode.
+    #[must_use]
+    #[doc(alias = "--dry-run")]
+    #[doc(alias = "-n")]
+    pub const fn dry_run(mut self, dry_run: bool) -> Self {
+        self.dry_run = dry_run;
+        self
+    }
+
+    /// Enables or disables list-only mode, mirroring `--list-only`.
+    #[must_use]
+    #[doc(alias = "--list-only")]
+    pub const fn list_only(mut self, list_only: bool) -> Self {
+        self.list_only = list_only;
+        self
+    }
+}
