+++ crates/core/src/client/config/builder/filters.rs
@@ -1,0 +1,32 @@
+use super::*;
+
+impl ClientConfigBuilder {
+    /// Replaces the collected debug flags with the provided list.
+    #[must_use]
+    #[doc(alias = "--debug")]
+    pub fn debug_flags<I, S>(mut self, flags: I) -> Self
+    where
+        I: IntoIterator<Item = S>,
+        S: Into<OsString>,
+    {
+        self.debug_flags = flags.into_iter().map(Into::into).collect();
+        self
+    }
+
+    /// Appends a filter rule to the configuration being constructed.
+    #[must_use]
+    pub fn add_filter_rule(mut self, rule: FilterRuleSpec) -> Self {
+        self.filter_rules.push(rule);
+        self
+    }
+
+    /// Extends the builder with a collection of filter rules.
+    #[must_use]
+    pub fn extend_filter_rules<I>(mut self, rules: I) -> Self
+    where
+        I: IntoIterator<Item = FilterRuleSpec>,
+    {
+        self.filter_rules.extend(rules);
+        self
+    }
+}
