+++ crates/core/src/client/config/builder/output.rs
@@ -1,0 +1,37 @@
+use super::*;
+
+impl ClientConfigBuilder {
+    /// Sets the verbosity level requested by the caller.
+    #[must_use]
+    #[doc(alias = "--verbose")]
+    #[doc(alias = "-v")]
+    pub const fn verbosity(mut self, verbosity: u8) -> Self {
+        self.verbosity = verbosity;
+        self
+    }
+
+    /// Enables or disables progress reporting for the transfer.
+    #[must_use]
+    #[doc(alias = "--progress")]
+    #[doc(alias = "--no-progress")]
+    pub const fn progress(mut self, progress: bool) -> Self {
+        self.progress = progress;
+        self
+    }
+
+    /// Enables or disables statistics reporting for the transfer.
+    #[must_use]
+    #[doc(alias = "--stats")]
+    pub const fn stats(mut self, stats: bool) -> Self {
+        self.stats = stats;
+        self
+    }
+
+    /// Enables or disables human-readable output formatting.
+    #[must_use]
+    #[doc(alias = "--human-readable")]
+    pub const fn human_readable(mut self, enabled: bool) -> Self {
+        self.human_readable = enabled;
+        self
+    }
+}
