+++ crates/core/src/client/config/builder/performance.rs
@@ -1,0 +1,78 @@
+use super::*;
+
+impl ClientConfigBuilder {
+    /// Configures the optional bandwidth limit to apply during transfers.
+    #[must_use]
+    #[doc(alias = "--bwlimit")]
+    pub fn bandwidth_limit(mut self, limit: Option<BandwidthLimit>) -> Self {
+        self.bandwidth_limit = limit;
+        self
+    }
+
+    /// Enables or disables compression for the transfer.
+    #[must_use]
+    #[doc(alias = "--compress")]
+    #[doc(alias = "--no-compress")]
+    #[doc(alias = "-z")]
+    pub const fn compress(mut self, compress: bool) -> Self {
+        self.compress = compress;
+        if compress && self.compression_setting.is_disabled() {
+            self.compression_setting = CompressionSetting::level(CompressionLevel::Default);
+        } else {
+            self.compression_setting = CompressionSetting::disabled();
+            self.compression_level = None;
+        }
+        self
+    }
+
+    /// Applies an explicit compression level override when building the configuration.
+    #[must_use]
+    #[doc(alias = "--compress-level")]
+    pub const fn compression_level(mut self, level: Option<CompressionLevel>) -> Self {
+        self.compression_level = level;
+        if let Some(level) = level {
+            self.compression_setting = CompressionSetting::level(level);
+            self.compress = true;
+        }
+        self
+    }
+
+    /// Sets the compression level that should apply when compression is enabled.
+    #[must_use]
+    #[doc(alias = "--compress-level")]
+    pub const fn compression_setting(mut self, setting: CompressionSetting) -> Self {
+        self.compression_setting = setting;
+        self.compress = setting.is_enabled();
+        if !self.compress {
+            self.compression_level = None;
+        }
+        self
+    }
+
+    /// Overrides the suffix list used to disable compression for specific extensions.
+    #[must_use]
+    #[doc(alias = "--skip-compress")]
+    pub fn skip_compress(mut self, list: SkipCompressList) -> Self {
+        self.skip_compress = list;
+        self
+    }
+
+    /// Requests that whole-file transfers be used instead of the delta algorithm.
+    #[must_use]
+    #[doc(alias = "--whole-file")]
+    #[doc(alias = "-W")]
+    #[doc(alias = "--no-whole-file")]
+    pub fn whole_file(mut self, whole_file: bool) -> Self {
+        self.whole_file = Some(whole_file);
+        self
+    }
+
+    /// Enables or disables sparse file handling for the transfer.
+    #[must_use]
+    #[doc(alias = "--sparse")]
+    #[doc(alias = "-S")]
+    pub const fn sparse(mut self, sparse: bool) -> Self {
+        self.sparse = sparse;
+        self
+    }
+}
