+++ crates/core/src/client/config/client/deletion.rs
@@ -1,0 +1,59 @@
+use super::*;
+
+impl ClientConfig {
+    /// Returns whether the run should avoid mutating the destination filesystem.
+    #[must_use]
+    #[doc(alias = "--dry-run")]
+    #[doc(alias = "-n")]
+    pub const fn dry_run(&self) -> bool {
+        self.dry_run
+    }
+
+    /// Returns the configured deletion mode.
+    #[must_use]
+    pub const fn delete_mode(&self) -> DeleteMode {
+        self.delete_mode
+    }
+
+    /// Returns whether extraneous destination files should be removed.
+    #[must_use]
+    #[doc(alias = "--delete")]
+    pub const fn delete(&self) -> bool {
+        self.delete_mode.is_enabled()
+    }
+
+    /// Returns whether extraneous entries should be removed before the transfer begins.
+    #[must_use]
+    #[doc(alias = "--delete-before")]
+    pub const fn delete_before(&self) -> bool {
+        matches!(self.delete_mode, DeleteMode::Before)
+    }
+
+    /// Returns whether extraneous entries should be removed after the transfer completes.
+    #[must_use]
+    #[doc(alias = "--delete-after")]
+    pub const fn delete_after(&self) -> bool {
+        matches!(self.delete_mode, DeleteMode::After)
+    }
+
+    /// Returns whether extraneous entries should be removed after transfers using delayed sweeps.
+    #[must_use]
+    #[doc(alias = "--delete-delay")]
+    pub const fn delete_delay(&self) -> bool {
+        matches!(self.delete_mode, DeleteMode::Delay)
+    }
+
+    /// Returns whether excluded destination entries should also be deleted.
+    #[must_use]
+    #[doc(alias = "--delete-excluded")]
+    pub const fn delete_excluded(&self) -> bool {
+        self.delete_excluded
+    }
+
+    /// Returns the configured maximum number of deletions, if any.
+    #[must_use]
+    #[doc(alias = "--max-delete")]
+    pub const fn max_delete(&self) -> Option<u64> {
+        self.max_delete
+    }
+}
